use usize instead of u64 for gas and priority

+++ coordinator/src/validator.rs
@@ -75,12 +75,12 @@ pub struct TransactionExecutionOutcome {
 
 pub enum TransactionCheckOutcome {
     IncludeInCurrentQueue {
-        priority: u64,
-        gas: u64,
+        priority: usize,
+        gas: usize,
     },
     IncludeInFutureQueue {
-        priority: u64,
-        gas: u64,
+        priority: usize,
+        gas: usize,
     },
     Exclude,
 }
